Removed privateSharing protocol, moved the relevant parts to privateChannelProtocolTest
