+++ src/helpers/assertEquals.ts
@@ -5,9 +5,7 @@ const defaultAreEqual = <T>(expected: T, actual: T): boolean => {
 export const assertEquals = <T>(expected: T, actual: T, areEqual = defaultAreEqual) => {
     if (areEqual(expected, actual) === false) {
         // tslint:disable-next-line:no-console
-        console.log('expected: ', expected);
-        // tslint:disable-next-line:no-console
-        console.log('actual: ', actual);
+        console.log('expected:', expected, '\nactual:', actual);
         throw new Error(`assertEquals failed: expected: ${expected}, actual: ${actual}`);
     }
 };
