+++ test/protocols/privateChannelTest.ts
@@ -3,8 +3,9 @@ import { syncPrivateChannelWithContact, privateChannelAddPost, makeEmptyPrivateC
 import { HexString } from '../../src/helpers/opaqueTypes';
 import { makeStorage, makeCrypto } from '../../src/cli/protocolTest/protocolTestHelpers';
 import { createAsyncDeterministicRandomGenerator } from '../../src/helpers/unsecureRandom';
-import { makePost } from '../../src/protocols/privateSharingTestHelpers';
 import { ImageData } from '../../src/models/ImageData';
+import { privateChannelProtocolTests } from '../../src/protocols/privateChannelProtocolTest';
+import { makePost } from '../../src/protocols/privateChannelProtocolTestHelpers';
 
 const testContactIdentity = {
     publicKey: '0x04d878f63e880d40ab684797469d38f7006c773a507624e4ec7a0cbf473bd52b4949a65ba56330a07647e0f0a2f7dd1d13cbe05c76206d532888f55fa79c51c41a',
@@ -25,6 +26,15 @@ const testMutualContact: MutualContact = {
 };
 
 const uploadImage = (image: ImageData): Promise<ImageData> => Promise.resolve(image);
+
+describe('Test private sharing', () => {
+    const tests: any = privateChannelProtocolTests;
+    for (const privateSharingTest of Object.keys(tests)) {
+        test('' + privateSharingTest, async () => {
+            await tests[privateSharingTest]();
+        });
+    }
+});
 
 describe('uploadUnsyncedCommands', () => {
 });
